Create README.md
Note: Two parameters are used to describe the structure ( metallization) in  systems (nano and microelectric and mechanical MEMS) . One is Ch, which starts from the main atom of the chiralized vector and is directed towards the next atom, and it   is also called  the metallization vector , and its length is equivalent to the environment of (nano, micro , electrical and mechanical MEMS)   systems   .
Chiral systems (nano and microelectromechanical MEMS)  are optically active and can have two non-equivalent spiral orientations. The process  ofmetallization  in  systems (nano, microelectric and mechanical MEMS) can be done both from above and below the graphene sheet. These two tubes will be mirror images of each other, and in this case, the graphene plane is the mirror plane. These two states create left-handed (M) and right-handed (P) isomers. In the case of chiralmetallization , these planes are called reflection planes, and both structures are equivalent in systems (nano and microelectric and mechanical MEMS)